# src/Backlog.App/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Backlog",
    "description": "Agile Software",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "Quinntyne Brown",
      "email": "quinntynebrown@gmail.com"
    },
    "license": {
      "name": "Use under MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/User/{userId}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "GetUserById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "UserId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetUserByIdResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "operationId": "RemoveUser",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "UserId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RemoveUserResponse"
            }
          }
        }
      }
    },
    "/api/User": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "GetUsers",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetUsersResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "CreateUser",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateUserRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CreateUserResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "operationId": "UpdateUser",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateUserRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UpdateUserResponse"
            }
          }
        }
      }
    },
    "/api/User/page/{pageSize}/{index}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "GetUsersPage",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "PageSize",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "Index",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetUsersPageResponse"
            }
          }
        }
      }
    },
    "/api/User/token": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "Authenticate",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AuthenticateRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AuthenticateResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "AuthenticateRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "AuthenticateResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string"
        },
        "userId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
...
    "CreateUserRequest": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/UserDto"
        }
      }
    },
    "CreateUserResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "user": {
          "$ref": "#/definitions/UserDto"
        }
      }
    },
...
    "GetUserByIdResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "user": {
          "$ref": "#/definitions/UserDto"
        }
      }
    },
    "GetUsersPageResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "length": {
          "format": "int32",
          "type": "integer"
        },
        "entities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserDto"
          }
        }
      }
    },
    "GetUsersResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserDto"
          }
        }
      }
    },
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
...
    "RemoveUserResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "user": {
          "$ref": "#/definitions/UserDto"
        }
      }
    },
...
    "UpdateUserRequest": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/UserDto"
        }
      }
    },
    "UpdateUserResponse": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "user": {
          "$ref": "#/definitions/UserDto"
        }
      }
    },
...
    "UserDto": {
      "type": "object",
      "properties": {
        "userId": {
          "format": "uuid",
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    }
  }
}